+++ src/index.ts
@@ -11,3 +11,30 @@ export * from "./table"
 export * from "./StopWatch"
 export * from "./strings"
 export * from "./BigDecimal"
+export { AsyncChainBase } from "./collections/async/AsyncChainBase"
+export { AsyncChain } from "./collections/async/AsyncChain"
+export { AsyncObjectChain } from "./collections/async/AsyncObjectChain"
+export { ValuesAsyncChain } from "./collections/async/ValuesAsyncChain"
+export { KeysAsyncChain } from "./collections/async/KeysAsyncChain"
+export { EntriesAsyncChain } from "./collections/async/EntriesAsyncChain"
+export { MappingKeysAsyncObjectChain } from "./collections/async/MappingKeysAsyncObjectChain"
+export { MappingValuesAsyncObjectChain } from "./collections/async/MappingValuesAsyncObjectChain"
+export { MappingEntriesAsyncObjectChain } from "./collections/async/MappingEntriesAsyncObjectChain"
+export { TakeWhileAsyncChain } from "./collections/async/TakeWhileAsyncChain"
+export { DropWhileAsyncChain } from "./collections/async/DropWhileAsyncChain"
+export { DistinctAsyncChain } from "./collections/async/DistinctAsyncChain"
+export { SliceAsyncChain } from "./collections/async/SliceAsyncChain"
+export { FilterAsyncChain } from "./collections/async/FilterAsyncChain"
+export { FlattenAsyncChain } from "./collections/async/FlattenAsyncChain"
+export { MappingAsyncChain } from "./collections/async/MappingAsyncChain"
+export { SortByAsyncChain } from "./collections/async/SortByAsyncChain"
+export { ZippingAsyncChain } from "./collections/async/ZippingAsyncChain"
+export { UnionAsyncChain } from "./collections/async/UnionAsyncChain"
+export { RunningReduceAsyncChain } from "./collections/async/RunningReduceAsyncChain"
+export { ConcatenatingAsyncChain } from "./collections/async/ConcatenatingAsyncChain"
+export { ChunkingAsyncChain } from "./collections/async/ChunkingAsyncChain"
+export { IntersectionAsyncChain } from "./collections/async/IntersectionAsyncChain"
+export { SlidingWindowAsyncChain } from "./collections/async/SlidingWindowAsyncChain"
+export { PermutationsAsyncChain } from "./collections/async/PermutationsAsyncChain"
+export { ReversingAsyncChain } from "./collections/async/ReversingAsyncChain"
+export { SortingAsyncChain } from "./collections/async/SortingAsyncChain"
